Improved how the board svg is rendered
Viewbox is now 0 to 1000 instead of having extra padding
The grid is rendered with two paths instead of a bunch of rects with stroke.
The svg now has a margin so it doesn't render at the very top-left of the screen.

Added ComponentRenderOnce and ComponentDirtyRender which can be inherited from to gain the common pattern of controlling exactly when to render.
 - All components that were individually copy/pasting this functionality now gain it through inheritance